Fix getZoneIDForEndpoint to properly calculate the longest match
The function was trying to find the longest match, but then confused the matching zoneID with the matching
zoneName. As zoneIDs are all of the same length it would pick the last zone that had a name longer than this
standard length (or the first zone, if there was never any long zone).